Extract helper from Layer to utils as `computeViewportFromCameraMatrix`
